Improved the way subpass dependencies are declared in AZ::Vulkan::RenderPass::ConvertRenderAttachmentLayout
To reduce complexity of AZ::Vulkan::RenderPass::ConvertRenderAttachmentLayout
and improve its readability added a new helper class named
SubpassDependencyHelper, which follows the same overarching algorithm
as the one done by the FrameGraphCompiler and Scope::Barrier management.

Added function to merge subpass dependencies:
Vulkan::RenderPass::Descriptor::MergeSubpassDependencies()
It was necessary to avoid validation errors when Vulkan compares
the VkRenderPass of the PipelineStateObject vs the VkRenderPass of the VkCmdBeginRenderPass.
Even if the subpass dependencies are identical but they differ in order, it'd trigger validation errors.
To make the order irrelevant, the solution is to merge the bitflags.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>